requirements.txt
asgiref==3.3.4
autopep8==1.5.7
bcrypt==3.2.2
beautifulsoup4==4.12.2
blinker==1.5
boto3==1.18.0
botocore==1.21.0
Brlapi==0.8.4
bs4==0.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.8.0
distro-info==1.5
Django==3.2.7
django-ckeditor==6.5.1
django-cors-headers==3.7.0
django-currencies==0.10.1
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
duplicity==0.8.22
fasteners==0.17.3
filelock==3.9.0
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
insta-scrape==2.1.2
instascrape==0.1
jeepney==0.8.0
jmespath==0.10.0
jsonfield==2.1.1
keyring==23.8.2
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.24.0
lxml==4.9.2
Mako==1.2.4.dev0
MarkupSafe==2.0.1
marshmallow==3.19.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==23.0
PAM==0.4.2
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==3.0.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.7.0
pycups==2.0.1
PyGObject==3.42.2
PyICU==2.9
PyJWT==2.1.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-apt==2.4.0+22.10
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.0
python-gettext==5.0
pytube==12.1.3
pytz==2021.1
pyxdg==0.28
PyYAML==5.4.1
reportlab==3.6.11
requests==2.28.2
s3transfer==0.5.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.4
sqlparse==0.4.1
systemd-python==235
toml==0.10.2
tomli==2.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.16
validate-email==1.3
virtualenv==20.19.0+ds
wadllib==1.3.6
whitenoise==5.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
